\name{xhmmScripts-package}
\alias{xhmmScripts-package}
\alias{xhmmScripts}
\docType{package}
\title{
Examine and plot XHMM whole-exome-sequencing-based CNV calls
}
\description{
  Examine and plot XHMM whole-exome-sequencing-based CNV
  calls.

  XHMM (eXome Hidden Markov Model) is a C++ software package
  (http://atgu.mgh.harvard.edu/xhmm) written to call copy number
  variation (CNV) from next-generation sequencing projects, where exome
  capture was used (or targeted sequencing, more generally).

  This R package enables the user to visualize both the PCA normalization
  performed by XHMM and the CNVs it has called.
}
\details{
\tabular{ll}{
Package: \tab xhmmScripts\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-22\cr
License: \tab GPL-2\cr
}
A typical use case is outlined here:\cr
https://bitbucket.org/statgen/xhmm/src/master/sources/scripts/example_make_XHMM_plots.R
}
\author{
  Menachem Fromer
  
Maintainer: Menachem Fromer <fromer@broadinstitute.org>
}
\references{
http://atgu.mgh.harvard.edu/xhmm/citations.shtml
}
